requirements.txt
pandas
fastparquet
matplotlib
isodate
gitpython